# semantics/framework/src/main/resources/static/aas-proxy-openapi.yaml
...
openapi: 3.0.3
info:
  title: DotAAS Part 2 | HTTP/REST | Entire Interface Collection
  description: The entire interface collection as part of Details of the Asset Administration Shell Part 2
  termsOfService: https://github.com/admin-shell-io/aas-specs
  contact:
    name: Michael Hoffmeister, Torben Miny, Andreas Orzelski, Manuel Sauer, Constantin Ziesche
  version: Final-Draft
servers:
- url: /
paths:
  /{assetIdentifier}/aas/submodels/{submodelIdentifier}:
    get:
      tags:
      - Submodel Repository Interface
      summary: Returns a specific Submodel
      operationId: GetSubmodelById
      parameters:
      - name: assetIdentifier
        in: path
        description: The Asset Administration Shells’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Requested Submodel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submodel'
      x-semanticIds:
      - https://admin-shell.io/aas/API/GetSubmodelById/1/0/RC02
    put:
      tags:
      - Submodel Repository Interface
      summary: Creates a new or updates an existing Submodel
      operationId: PutSubmodelById
      parameters:
      - name: assetIdentifier
        in: path
        description: The Asset Administration Shells’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Submodel object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Submodel'
        required: true
      responses:
        "201":
          description: Submodel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submodel'
        "204":
          description: Submodel updated successfully
      x-semanticIds:
      - https://admin-shell.io/aas/API/PutSubmodelById/1/0/RC02
    delete:
      tags:
      - Submodel Repository Interface
      summary: Deletes a Submodel
      operationId: DeleteSubmodelById
      parameters:
      - name: assetIdentifier
        in: path
        description: The Asset Administration Shells’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Submodel deleted successfully
      x-semanticIds:
      - https://admin-shell.io/aas/API/DeleteSubmodelById/1/0/RC02
  /{assetIdentifier}/aas/{submodelIdentifier}/submodel:
    get:
      tags:
      - Submodel Interface
      summary: Returns the Submodel
      operationId: GetSubmodel
      parameters:
      - name: assetIdentifier
        in: path
        description: The Asset Administration Shells’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: level
        in: query
        description: Determines the structural depth of the respective resource content
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: deep
          enum:
          - deep
          - core
      - name: content
        in: query
        description: Determines the request or response kind of the resource
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: normal
          enum:
          - normal
          - trimmed
          - value
          - reference
          - path
      - name: extent
        in: query
        description: Determines to which extent the resource is being serialized
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - withBlobValue
          - withoutBlobValue
      responses:
        "200":
          description: Requested Submodel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submodel'
      x-semanticIds:
      - https://admin-shell.io/aas/API/GetSubmodel/1/0/RC02
    put:
      tags:
      - Submodel Interface
      summary: Updates the Submodel
      operationId: PutSubmodel
      parameters:
      - name: assetIdentifier
        in: path
        description: The Asset Administration Shells’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: level
        in: query
        description: Determines the structural depth of the respective resource content
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: deep
          enum:
          - deep
          - core
      - name: content
        in: query
        description: Determines the request or response kind of the resource
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: normal
          enum:
          - normal
          - trimmed
          - value
          - reference
          - path
      - name: extent
        in: query
        description: Determines to which extent the resource is being serialized
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - withBlobValue
          - withoutBlobValue
      requestBody:
        description: Submodel object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Submodel'
        required: true
      responses:
        "204":
          description: Submodel updated successfully
      x-semanticIds:
      - https://admin-shell.io/aas/API/PutSubmodel/1/0/RC02
  /{assetIdentifier}/aas/{submodelIdentifier}/submodel/submodel-elements:
    get:
      tags:
      - Submodel Interface
      summary: Returns all submodel elements including their hierarchy
      operationId: GetAllSubmodelElements
      parameters:
      - name: assetIdentifier
        in: path
        description: The Asset Administration Shells’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: level
        in: query
        description: Determines the structural depth of the respective resource content
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: deep
          enum:
          - deep
          - core
      - name: content
        in: query
        description: Determines the request or response kind of the resource
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: normal
          enum:
          - normal
          - trimmed
          - value
          - reference
          - path
      - name: extent
        in: query
        description: Determines to which extent the resource is being serialized
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - withBlobValue
          - withoutBlobValue
      responses:
        "200":
          description: List of found submodel elements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubmodelElement'
      x-semanticIds:
      - https://admin-shell.io/aas/API/GetAllSubmodelElements/1/0/RC02
      - https://admin-shell.io/aas/API/GetAllSubmodelElementsBySemanticId/1/0/RC02
      - https://admin-shell.io/aas/API/GetAllSubmodelElementsByParentPathAndSemanticId/1/0/RC02
    post:
      tags:
      - Submodel Interface
      summary: Creates a new submodel element
      operationId: PostSubmodelElement
      parameters:
      - name: assetIdentifier
        in: path
        description: The Asset Administration Shells’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: level
        in: query
        description: Determines the structural depth of the respective resource content
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: deep
          enum:
          - deep
          - core
      - name: content
        in: query
        description: Determines the request or response kind of the resource
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: normal
          enum:
          - normal
          - trimmed
          - value
          - reference
          - path
      - name: extent
        in: query
        description: Determines to which extent the resource is being serialized
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - withBlobValue
          - withoutBlobValue
      requestBody:
        description: Requested submodel element
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmodelElement'
        required: true
      responses:
        "201":
          description: Submodel element created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmodelElement'
      x-semanticIds:
      - https://admin-shell.io/aas/API/PostSubmodelElement/1/0/RC02c
...
  /{assetIdentifier}/aas/{submodelIdentifier}/submodel/submodel-elements/{idShortPath}/invoke:
    post:
      tags:
      - Submodel Interface
      summary: Synchronously or asynchronously invokes an Operation at a specified path
      operationId: InvokeOperation
      parameters:
      - name: assetIdentifier
        in: path
        description: The Asset Administration Shells’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: async
        in: query
        description: Determines whether an operation invocation is performed asynchronously or synchronously
        required: false
        style: form
        explode: true
        schema:
          type: boolean
          default: false
      - name: content
        in: query
        description: Determines the request or response kind of the resource
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: normal
          enum:
          - normal
          - trimmed
          - value
          - reference
          - path
      - name: idShortPath
        in: path
        description: IdShort path to the submodel element (dot-separated), in this case an operation
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Operation request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationRequest'
        required: true
      responses:
        "200":
          description: Operation result object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResult'
      x-semanticIds:
      - https://admin-shell.io/aas/API/InvokeOperationSync/1/0/RC02
      - https://admin-shell.io/aas/API/InvokeOperationAsync/1/0/RC02
  /{assetIdentifier}/aas/{submodelIdentifier}/submodel/submodel-elements/{idShortPath}/operation-results/{handleId}:
    get:
      tags:
      - Submodel Interface
      summary: Returns the Operation result of an asynchronous invoked Operation
      operationId: GetOperationAsyncResult
      parameters:
      - name: assetIdentifier
        in: path
        description: The Asset Administration Shells’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: submodelIdentifier
        in: path
        description: The Submodel’s unique id (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: content
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: normal
          enum:
          - normal
          - value
      - name: idShortPath
        in: path
        description: IdShort path to the submodel element (dot-separated), in this case an operation
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: handleId
        in: path
        description: The returned handle id of an operation’s asynchronous invocation used to request the current state of the operation’s execution (BASE64-URL-encoded)
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Operation result object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResult'
      x-semanticIds:
      - https://admin-shell.io/aas/API/GetOperationAsyncResult/1/0/RC02
components:
  schemas:
...
    HasSemantics:
      type: object
      properties:
        semanticId:
          $ref: '#/components/schemas/Reference'
...
    Identifiable:
      allOf:
      - $ref: '#/components/schemas/Referable'
      - required:
        - identification
        properties:
          administration:
            $ref: '#/components/schemas/AdministrativeInformation'
          identification:
            $ref: '#/components/schemas/Identifier'
    Referable:
      allOf:
      - $ref: '#/components/schemas/HasExtensions'
      - required:
        - idShort
        - modelType
        properties:
          category:
            type: string
          description:
            type: array
            items:
              $ref: '#/components/schemas/LangString'
          displayName:
            type: array
            items:
              $ref: '#/components/schemas/LangString'
          idShort:
            type: string
          modelType:
            $ref: '#/components/schemas/ModelType'
...
    HasDataSpecification:
      type: object
      properties:
        embeddedDataSpecifications:
          type: array
          items:
            $ref: '#/components/schemas/EmbeddedDataSpecification'
...
    File:
      allOf:
      - $ref: '#/components/schemas/SubmodelElement'
      - required:
        - mimeType
        properties:
          mimeType:
            type: string
          value:
            type: string
    SubmodelElement:
      allOf:
      - $ref: '#/components/schemas/Referable'
      - $ref: '#/components/schemas/HasDataSpecification'
      - $ref: '#/components/schemas/HasSemantics'
      - $ref: '#/components/schemas/Qualifiable'
      - properties:
          kind:
            $ref: '#/components/schemas/ModelingKind'
      oneOf:
      - $ref: '#/components/schemas/Blob'
      - $ref: '#/components/schemas/File'
      - $ref: '#/components/schemas/Capability'
      - $ref: '#/components/schemas/Entity'
      - $ref: '#/components/schemas/Event'
      - $ref: '#/components/schemas/BasicEvent'
      - $ref: '#/components/schemas/MultiLanguageProperty'
      - $ref: '#/components/schemas/Operation'
      - $ref: '#/components/schemas/Property'
      - $ref: '#/components/schemas/Range'
      - $ref: '#/components/schemas/ReferenceElement'
      - $ref: '#/components/schemas/AnnotatedRelationshipElement'
      - $ref: '#/components/schemas/RelationshipElement'
      - $ref: '#/components/schemas/SubmodelElementList'
      - $ref: '#/components/schemas/SubmodelElementStruct'
    Qualifiable:
      type: object
      properties:
        qualifiers:
          type: array
          items:
            $ref: '#/components/schemas/Constraint'
...
    ModelingKind:
      type: string
      enum:
      - Template
      - Instance
    Blob:
      allOf:
      - $ref: '#/components/schemas/SubmodelElement'
      - required:
        - mimeType
        properties:
          mimeType:
            type: string
          value:
            type: string
    Capability:
      allOf:
      - $ref: '#/components/schemas/Referable'
      - $ref: '#/components/schemas/HasDataSpecification'
      - $ref: '#/components/schemas/HasSemantics'
      - $ref: '#/components/schemas/Qualifiable'
      - properties:
          kind:
            $ref: '#/components/schemas/ModelingKind'
      oneOf:
      - $ref: '#/components/schemas/Blob'
      - $ref: '#/components/schemas/File'
      - $ref: '#/components/schemas/Capability'
      - $ref: '#/components/schemas/Entity'
      - $ref: '#/components/schemas/Event'
      - $ref: '#/components/schemas/BasicEvent'
      - $ref: '#/components/schemas/MultiLanguageProperty'
      - $ref: '#/components/schemas/Operation'
      - $ref: '#/components/schemas/Property'
      - $ref: '#/components/schemas/Range'
      - $ref: '#/components/schemas/ReferenceElement'
      - $ref: '#/components/schemas/AnnotatedRelationshipElement'
      - $ref: '#/components/schemas/RelationshipElement'
      - $ref: '#/components/schemas/SubmodelElementList'
      - $ref: '#/components/schemas/SubmodelElementStruct'
...
    Event:
      allOf:
      - $ref: '#/components/schemas/SubmodelElement'
...
    Property:
      allOf:
      - $ref: '#/components/schemas/SubmodelElement'
      - $ref: '#/components/schemas/ValueObject'
...
    Submodel:
      allOf:
      - $ref: '#/components/schemas/Identifiable'
      - $ref: '#/components/schemas/HasDataSpecification'
      - $ref: '#/components/schemas/Qualifiable'
      - $ref: '#/components/schemas/HasSemantics'
      - properties:
          kind:
            $ref: '#/components/schemas/ModelingKind'
          submodelElements:
            type: array
            items:
              $ref: '#/components/schemas/SubmodelElement'
    OperationRequest:
      type: object
      properties:
        inoutputArguments:
          type: array
          items:
            $ref: '#/components/schemas/OperationVariable'
        inputArguments:
          type: array
          items:
            $ref: '#/components/schemas/OperationVariable'
        requestId:
          type: string
        timeout:
          type: integer
    OperationResult:
      type: object
      properties:
        executionResult:
          $ref: '#/components/schemas/Result'
        executionState:
          type: string
          enum:
          - Initiated
          - Running
          - Completed
          - Canceled
          - Failed
          - Timeout
        inoutputArguments:
          type: array
          items:
            $ref: '#/components/schemas/OperationVariable'
        outputArguments:
          type: array
          items:
            $ref: '#/components/schemas/OperationVariable'
        requestId:
          type: string
...